+++ packages/arguments-builder/jest.config.js
@@ -1,7 +1,10 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} **/
 module.exports = {
-  testEnvironment: "node",
+  testEnvironment: 'node',
   transform: {
-    "^.+.tsx?$": ["ts-jest",{}],
+    '^.+.tsx?$': ['ts-jest', {}],
+  },
+  moduleNameMapper: {
+    '@iringo/arguments-builder': '<rootDir>/src/index.ts',
   },
 };
