+++ packages/arguments-builder/rslib.config.ts
@@ -1,6 +1,12 @@
 import { defineConfig } from '@rslib/core';
 
 export default defineConfig({
+  source: {
+    entry: {
+      index: './src/index.ts',
+      bin: './src/bin.ts',
+    },
+  },
   lib: [
     {
       format: 'esm',
