+++ packages/arguments-builder/src/index.ts
@@ -1,2 +1,8 @@
+import type { ArgumentsBuilderConfig } from './bin';
+
 export * from './core';
 export * from './core/types';
+
+export function defineConfig(config: ArgumentsBuilderConfig) {
+  return config;
+}
