+++ packages/arguments-builder/tests/config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from '../src';
+import { testArgs } from './common';
+
+export default defineConfig({
+  args: testArgs,
+});
